Refactor route
