+++ .babelrc
@@ -1,0 +1,12 @@
+
+{
+    "presets": ["@babel/preset-react", "@babel/preset-env"],
+    "plugins": [
+      [
+        "module-resolver",
+        {
+            "root": ["./src"]
+        }
+      ]
+    ]
+  }
