+++ package.json
@@ -57,7 +57,9 @@
     ]
   },
   "devDependencies": {
+    "@babel/preset-react": "^7.18.6",
     "autoprefixer": "^10.4.13",
+    "babel-plugin-module-resolver": "^5.0.0",
     "jest": "^27.5.1",
     "postcss": "^8.4.19",
     "tailwindcss": "^3.2.4"
