+++ src/components/index.jsx
@@ -1,3 +1,1 @@
-import Footer from './footer'
-
-export {Footer}
+export * from './footer'
