--- src/pages/contactUs/index.jsx
@@ -1,11 +1,0 @@
-import ContactUsComp from "./contactUsComp/ContactUsComp";
-
-const ContactUs = () => {
-  return (
-    <>
-      <ContactUsComp />
-    </>
-  );
-};
-
-export default ContactUs;
