+++ src/pages/external/contactUs/components/index.jsx
@@ -1,0 +1,2 @@
+export * from './contactUsComp'
+
