+++ src/pages/external/contactUs/index.jsx
@@ -1,0 +1,10 @@
+import { ContactUsComp } from "./components";
+
+export const ContactUs = () => {
+  return (
+    <>
+      <ContactUsComp />
+    </>
+  );
+};
+
