+++ src/pages/external/index.jsx
@@ -1,0 +1,2 @@
+export * from './contactUs'
+export * from './policy'
