+++ src/pages/index.jsx
@@ -1,0 +1,1 @@
+export * from './external'
