+++ src/pages/t&c/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Navbar from "../../components/Navbar";
-import Footer from "../../components/footer/index";
+import {Footer} from "../../components";
 
 const Terms = () => {
   const contents = [
