+++ src/routes/public/publicRoutesData.js
@@ -4,8 +4,8 @@ import SignIn from "../../pages/signIn";
 import ForgetPassword from "../../pages/ForgetPassword/index";
 import ResetLink from "../../pages/ForgetPassword/LinkReset";
 import ResetPassword from "../../pages/ForgetPassword/ResetPassword";
-import ContactUs from '../../pages/contactUs';
-import Policy from "../../pages/external/policy";
+import { ContactUs } from '../../pages';
+import { Policy } from "../../pages";
 import HowItWorks from "../../pages/howItWorks/HowItWorks";
 import Terms from "../../pages/t&c";
 import TermsOfUse from "../../pages/termsOfUse";
